+++ packages/widgetbook/lib/src/addons/text_scale_addon/text_scale_addon.dart
@@ -1,42 +1,59 @@
-// [MediaQuery.textScaleFactor] is deprecated in Flutter 3.16.0,
-// Since our minimum Flutter version is 3.7.0, we cannot use [TextScaler] yet.
-// More info: https://docs.flutter.dev/release/breaking-changes/deprecate-textscalefactor
-// ignore_for_file: deprecated_member_use
-
 import 'package:flutter/material.dart';
 
 import '../../fields/fields.dart';
 import '../common/common.dart';
 
-/// A [WidgetbookAddon] for changing the active [MediaQueryData.textScaleFactor]
+/// A [WidgetbookAddon] for changing the active [MediaQueryData.textScaler]
 /// via [MediaQuery].
 class TextScaleAddon extends WidgetbookAddon<double> {
   TextScaleAddon({
-    required this.scales,
+    @Deprecated('Use TextScaleAddon.min and TextScaleAddon.max instead')
+    this.scales,
     this.initialScale,
+    this.min = 0.5,
+    this.max = 2.0,
+    this.divisions = 6,
   })  : assert(
-          scales.isNotEmpty,
-          'scales cannot be empty',
+          scales == null || scales.isNotEmpty,
+          'scales must not be empty, if set',
         ),
         assert(
-          initialScale == null || scales.contains(initialScale),
+          scales == null ||
+              initialScale == null ||
+              scales.contains(initialScale),
           'initialScale must be in scales',
         ),
         super(
           name: 'Text scale',
         );
 
+  final List<double>? scales;
   final double? initialScale;
-  final List<double> scales;
+  final double min;
+  final double max;
+  final int divisions;
 
   @override
   List<Field> get fields {
+    // Fallback to old implementation if scales are provided
+    if (scales != null) {
+      return [
+        ListField<double>(
+          name: 'factor',
+          values: scales!,
+          initialValue: initialScale ?? scales!.first,
+          labelBuilder: (scale) => scale.toStringAsFixed(2),
+        ),
+      ];
+    }
+
     return [
-      ListField<double>(
+      DoubleSliderField(
         name: 'factor',
-        values: scales,
-        initialValue: initialScale ?? scales.first,
-        labelBuilder: (scale) => scale.toStringAsFixed(2),
+        initialValue: initialScale ?? 1.0,
+        min: min,
+        max: max,
+        divisions: divisions,
       ),
     ];
   }
@@ -54,7 +71,7 @@ class TextScaleAddon extends WidgetbookAddon<double> {
   ) {
     return MediaQuery(
       data: MediaQuery.of(context).copyWith(
-        textScaleFactor: setting,
+        textScaler: TextScaler.linear(setting),
       ),
       child: child,
     );