+++ packages/widgetbook/lib/src/fields/field_codec.dart
@@ -25,13 +25,33 @@ class FieldCodec<T> {
   /// print(encoded); // {foo:bar,baz:qux}
   /// ```
   static String encodeQueryGroup(Map<String, String> group) {
-    final pairs = group.entries.map((entry) => '${entry.key}:${entry.value}');
+    final pairs = group.entries.map((entry) {
+      // Both key and value are encoded to ensure that reserved
+      // characters (e.g. `:`, `{`, `}` and `,`) are not misinterpreted.
+      // For example, using a comma in a string value or a colon
+      // in a date value would break the decoding process.
+      final encodedKey = Uri.encodeComponent(entry.key);
+      final encodedValue = Uri.encodeComponent(entry.value);
+
+      return '$encodedKey:$encodedValue';
+    });
+
     return '{${pairs.join(',')}}';
+  }
+
+  /// Decodes [component] using [Uri.decodeComponent],
+  /// but returns null if the decoding fails due to non-ASCII characters.
+  static String? tryDecodeComponent(String component) {
+    try {
+      return Uri.decodeComponent(component);
+    } on ArgumentError {
+      return null;
+    }
   }
 
   /// Decodes a query group encoded value back to a [Map].
   static Map<String, String> decodeQueryGroup(String? group) {
-    if (group == null) return {};
+    if (group == null || group == '{}') return {};
 
     final params = group.substring(1, group.length - 1).split(',');
 
@@ -39,7 +59,10 @@ class FieldCodec<T> {
       params.map(
         (param) {
           final parts = param.split(':');
-          return MapEntry(parts[0], parts[1]);
+          final decodedKey = tryDecodeComponent(parts[0]);
+          final decodedValue = tryDecodeComponent(parts[1]);
+
+          return MapEntry(decodedKey ?? parts[0], decodedValue ?? parts[1]);
         },
       ),
     );