+++ packages/widgetbook/lib/src/state/widgetbook_state.dart
@@ -73,11 +73,19 @@ class WidgetbookState extends ChangeNotifier {
     );
   }
 
+  /// Gets the current state using [context], if any.
+  /// If there is no state in scope, then this function will return null.
+  static WidgetbookState? maybeOf(BuildContext context) {
+    return context
+        .dependOnInheritedWidgetOfExactType<WidgetbookScope>()
+        ?.notifier;
+  }
+
   /// Gets the current state using [context].
   static WidgetbookState of(BuildContext context) {
-    return context
-        .dependOnInheritedWidgetOfExactType<WidgetbookScope>()!
-        .notifier!;
+    final state = WidgetbookState.maybeOf(context);
+    assert(state != null, 'No Widgetbook found in the context.');
+    return state!;
   }
 
   @internal
@@ -99,9 +107,7 @@ class WidgetbookState extends ChangeNotifier {
   /// Syncs this with the router's location using [SystemNavigator].
   void _syncRouteInformation() {
     SystemNavigator.routeInformationUpdated(
-      // Not backwards compatible with Flutter < 3.13.0
-      // ignore: deprecated_member_use
-      location: uri.toString(),
+      uri: uri,
     );
   }