+++ packages/widgetbook/test/helper/tester_extension.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:widgetbook/src/state/widgetbook_scope.dart';
 import 'package:widgetbook/src/themes.dart';
 import 'package:widgetbook/widgetbook.dart';
 
@@ -86,13 +85,14 @@ extension TesterExtension on WidgetTester {
     Field<TValue> field,
     TValue? value,
   ) async {
-    const group = 'group_name';
+    const groupKey = 'group_name';
+    final groupValue = FieldCodec.encodeQueryGroup(
+      value != null ? {field.name: field.codec.toParam(value)} : {},
+    );
 
     await pumpWidgetWithQueryParams(
-      queryParams: value != null
-          ? {group: '{${field.name}:${field.codec.toParam(value)}}'}
-          : {},
-      builder: (context) => field.build(context, group),
+      queryParams: value != null ? {groupKey: groupValue} : {},
+      builder: (context) => field.build(context, groupKey),
     );
 
     return widget<TWidget>(