+++ packages/widgetbook_cli/lib/src/api/api.dart
@@ -1,5 +1,6 @@
-export 'models/build_request.dart';
-export 'models/build_response.dart';
-export 'models/upload_task.dart';
+export 'models/build_draft_request.dart';
+export 'models/build_draft_response.dart';
+export 'models/build_ready_request.dart';
+export 'models/build_ready_response.dart';
 export 'models/versions_metadata.dart';
 export 'widgetbook_http_client.dart';